Allow building under longer paths
It turns out that #41 was insufficient since `cmake` will forcibly _set_
`CMAKE_INSTALL_PREFIX` to the output directory (which is under the
target directory) if it isn't already set[1]. So, we need to explicitly
set a short dummy value.

This should hopefully _actually_ fix docs.rs builds.

The version bump to ibverbs (non-sys) is just to force docs.rs to
attempt re-generation of the docs.

[1]: https://github.com/rust-lang/cmake-rs/blob/94da9de2ea79ab6cad572e908864a160cf4847a9/src/lib.rs#L699-L703